+++ src/main/java/nl/esciencecenter/computeservice/service/XenonMonitor.java
@@ -20,7 +20,7 @@ import nl.esciencecenter.computeservice.service.tasks.CwlStageInTask;
 import nl.esciencecenter.computeservice.service.tasks.CwlStageOutTask;
 import nl.esciencecenter.computeservice.service.tasks.CwlWorkflowTask;
 import nl.esciencecenter.computeservice.service.tasks.DeleteJobTask;
-import nl.esciencecenter.computeservice.utils.RemoteFilesystemUtils;
+import nl.esciencecenter.computeservice.utils.FileSystemUtils;
 import nl.esciencecenter.xenon.XenonException;
 import nl.esciencecenter.xenon.adaptors.schedulers.JobCanceledException;
 import nl.esciencecenter.xenon.schedulers.JobStatus;
@@ -264,7 +264,7 @@ public class XenonMonitor {
 						
 						try {
 							// delete remote directory
-							RemoteFilesystemUtils.deleteRemoteWorkdir(xenonService.getRemoteFileSystem(), jobLogger, job);
+							FileSystemUtils.deleteRemoteWorkdir(xenonService.getRemoteFileSystem(), jobLogger, job);
 						} catch (XenonException e) {
 							jobLogger.info("Exception while deleting remote directory:", e);
 							logger.error("Exception while deleting remote directory:", e);