+++ src/test/java/nl/esciencecenter/computeservice/rest/CwlResultTest.java
@@ -7,6 +7,8 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import java.util.UUID;
+
 import org.apache.commons.io.IOUtils;
 import org.junit.After;
 import org.junit.Test;
@@ -23,9 +25,11 @@ import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
+import nl.esciencecenter.client.CWLState;
 import nl.esciencecenter.client.Job;
 import nl.esciencecenter.computeservice.service.XenonService;
 import nl.esciencecenter.computeservice.utils.CwlTestUtils;
+import nl.esciencecenter.xenon.filesystems.CopyMode;
 import nl.esciencecenter.xenon.filesystems.FileSystem;
 import nl.esciencecenter.xenon.filesystems.Path;
 
@@ -51,7 +55,7 @@ public class CwlResultTest {
 	@After
 	public void deleteJob() throws Exception {
 		for (Job job : CwlTestUtils.getCreated()) {
-			this.mockMvc.perform(
+			mockMvc.perform(
 					delete(job.getUri())
 					.header(headerName, apiToken)
 				);
@@ -71,7 +75,7 @@ public class CwlResultTest {
 			String contents = IOUtils.toString(this.getClass().getClassLoader().getResourceAsStream("jobs/echo-test.json"), "UTF-8");
 			Job job = CwlTestUtils.postJobAndWaitForFinal(contents, mockMvc, headerName, apiToken);
 			
-			MockHttpServletResponse response = this.mockMvc.perform(
+			MockHttpServletResponse response = mockMvc.perform(
 					get(job.getLog())
 					.header(headerName, apiToken)
 			).andExpect(status().is2xxSuccessful()).andReturn().getResponse();
@@ -86,7 +90,7 @@ public class CwlResultTest {
 			String contents = IOUtils.toString(this.getClass().getClassLoader().getResourceAsStream("jobs/jobid-test.json"), "UTF-8");
 			Job job = CwlTestUtils.postJobAndWaitForFinal(contents, mockMvc, headerName, apiToken);
 			
-			MockHttpServletResponse response = this.mockMvc.perform(
+			MockHttpServletResponse response = mockMvc.perform(
 					get(job.getLog())
 					.header(headerName, apiToken)
 			).andExpect(status().is2xxSuccessful()).andReturn().getResponse();
@@ -94,6 +98,38 @@ public class CwlResultTest {
 			
 			String output = (String) job.getOutput().get("out");
 			assertTrue(output.equals(job.getId() + " " + job.getName()+"\n"));
+		}).doesNotThrowAnyException();
+	}
+	
+	@Test
+	public void testDeletingInput() {
+		assertThatCode(() -> {
+			xenonService.getConfig().getSourceFilesystemConfig().setClearOnJobDone(true);
+			
+			// Create an input directory and populate it with an input file.
+			String uuid = UUID.randomUUID().toString();
+			FileSystem inputFileSystem = xenonService.getSourceFileSystem();
+			
+			Path inputDir = inputFileSystem.getWorkingDirectory().resolve(uuid);
+			inputFileSystem.createDirectory(inputDir);
+			
+			Path sourceFile = inputFileSystem.getWorkingDirectory().resolve("input").resolve("lorem.txt");
+			Path inputFile = inputDir.resolve("lorem.txt");
+			String copyId = inputFileSystem.copy(sourceFile, inputFileSystem, inputFile, CopyMode.IGNORE, false);
+			inputFileSystem.waitUntilDone(copyId, 1000);
+			
+			
+			String contents = IOUtils.toString(this.getClass().getClassLoader().getResourceAsStream("jobs/copy-test.json"), "UTF-8");
+			contents = contents.replace("input/", uuid+"/");
+			
+			Job job = CwlTestUtils.postJobAndWaitForFinal(contents, mockMvc, headerName, apiToken);
+			assertTrue(job.getState() == CWLState.SUCCESS);
+			
+			assertFalse(inputFileSystem.exists(inputFile));
+			
+			inputFileSystem.delete(inputDir, true);
+			
+			xenonService.getConfig().getSourceFilesystemConfig().setClearOnJobDone(false);
 		}).doesNotThrowAnyException();
 	}
 }